Commit initial hacked-up copy of "famms" template.
At present this is a static web site, and content
adaptation has only just begun. Incommon with many
designs the home page allows the reader to scroll
through the entire site contents, but this leads to
wasteful and error-prone duplication of content.

The first task will be to recast the existing site
as a correlated set of Jinja2 templates, into which
the site programming (Flask, if you are interested)
injects the relevant content.